rename metadata filter form files to match component names
date-metadata-filter-form.tsx and user-metadata-filter-form.tsx
