+++ frontend/app/src/entities/nodes/object/ui/filters/field-filter-form.tsx
@@ -1,7 +1,7 @@
 import type { FilterDefinition } from "@/entities/nodes/object/domain/filter-definition";
 import { AttributeFilterForm } from "@/entities/nodes/object/ui/filters/attribute-filter-form";
-import { DateMetadataFilterForm } from "@/entities/nodes/object/ui/filters/metadata-date-filter-form";
-import { UserMetadataFilterForm } from "@/entities/nodes/object/ui/filters/metadata-user-filter-form";
+import { DateMetadataFilterForm } from "@/entities/nodes/object/ui/filters/date-metadata-filter-form";
+import { UserMetadataFilterForm } from "@/entities/nodes/object/ui/filters/user-metadata-filter-form";
 import { RelationshipFilterForm } from "@/entities/nodes/object/ui/filters/relationship-filter-form";
 
 interface FieldFilterFormProps {
